// src/products/types/products.ts
export interface Product {
  Cod: number;
  Descripción: string;
  Descripción2: string;
  Marca: string;
  Cod_proveedor: number;
  Costo_unitario: number;
  Stock_total: number;
  igv: number;
  Valor_de_mercancia: number;
  Fecha_creación: string;
  Usuario: string;
  Ult_modificación: string;
  Usuario_modificador: string;
  Estado: boolean;
}